Separate out EmptyAreaCollection and ClusteredAreaCollection
