+++ viewer/packages/cad-styling/index.ts
@@ -9,5 +9,7 @@ export { UnionNodeCollection } from './src/UnionNodeCollection';
 export { NodeAppearance, DefaultNodeAppearance, NodeOutlineColor } from './src/NodeAppearance';
 export { NodeAppearanceProvider } from './src/NodeAppearanceProvider';
 
-export { AreaCollection, EmptyAreaCollection, ClusteredAreaCollection } from './src/prioritized/AreaCollection';
+export { AreaCollection } from './src/prioritized/AreaCollection';
+export { EmptyAreaCollection } from './src/prioritized/EmptyAreaCollection';
+export { ClusteredAreaCollection } from './src/prioritized/ClusteredAreaCollection';
 export { PrioritizedArea } from './src/prioritized/types';
